pyproject.toml
[tool.poetry]
name = "chemqulacs"
version = "0.1.0"
description = "Quantum chemistry code running on quantum circuit simulators and quantum computers."
authors = ["chemqulacs developers"]
license = "Apache-2.0"
repository = "https://github.com/wmizukami/chemqulacs"
documentation = "https://wmizukami.github.io/chemqulacs/"
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.9.8,<3.12"
pyscf = "^2.0.1"
openfermion = "^1.3.0"
numpy = "^1.22.3"
pubchempy = "^1.0.4"
scipy = "~1.13.1"

quri-parts-algo = "^0.19.0"
quri-parts-braket = "^0.19.0"
quri-parts-chem = "^0.19.0"
quri-parts-core = "^0.19.0"
quri-parts-circuit = "^0.19.0"
quri-parts-itensor = "^0.19.0"
quri-parts-qulacs = "^0.19.0"
quri-parts-openfermion = "^0.19.0"
quri-parts-qiskit = "^0.19.0"
qiskit-ibm-runtime = "^0.22.0"
urllib3 = "<2" 


[tool.poetry.group.dev.dependencies]
pytest = "^6.2.5"
black = "^22.3.0"
flake8 = "^4.0.1"
mypy = "^0.910"
isort = "^5.10.1"
Sphinx = ">=5.0.0"
sphinx-book-theme = "^1.0.0"
ipykernel = "^6.4.2"
myst-nb = "^0.17.2"
openfermionpyscf = "^0.5"

[tool.isort]
default_section = "THIRDPARTY"
ensure_newline_before_comments = true
force_grid_wrap = 0
force_single_line = false
include_trailing_comma = true
multi_line_output = 3
use_parentheses = true

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
